+++ EssentialFeed/EssentialFeedTests/Feed API/RemoteFeedImageDataLoaderTests.swift
@@ -8,59 +8,6 @@
 import XCTest
 import EssentialFeed
 
-final class RemoteFeedImageDataLoader {
-    let client : HTTPClient
-    
-    init(client: HTTPClient) {
-        self.client = client
-    }
-    
-    public enum Error: Swift.Error {
-        case invalidData
-    }
-    
-    private final class HTTPClientTaskWrapper: FeedImageDataLoaderTask {
-        private var completion: ((FeedImageDataLoader.Result) -> Void)?
-        var wrapped: HTTPClientTask?
-        
-        init(_ completion: @escaping (FeedImageDataLoader.Result) -> Void) {
-            self.completion = completion
-        }
-        
-        func complete(with result: FeedImageDataLoader.Result) {
-            completion?(result)
-        }
-        
-        func cancel() {
-            preventFurtherCompletion()
-            wrapped?.cancel()
-        }
-        
-        private func preventFurtherCompletion() {
-            completion = nil
-        }
-    }
-    
-    @discardableResult
-    func loadImageDataFromURL(from url: URL, completion:@escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask  {
-        let task = HTTPClientTaskWrapper(completion)
-        task.wrapped = client.get(from: url) { [weak self] result in
-            guard let _ = self else { return }
-            switch result {
-            case let .success((data, response)):
-                if response.statusCode == 200, data.isEmpty == false {
-                    task.complete(with: .success(data))
-                } else {
-                    task.complete(with: .failure(Error.invalidData))
-                }
-            case let .failure(error): task.complete(with: .failure(error))
-            }
-        }
-        
-        return task
-    }
-}
-
 class RemoteFeedImageDataLoaderTests: XCTestCase {
     
     func test_init_doesNotPerformAnyURLRequest() {
@@ -202,6 +149,4 @@ class RemoteFeedImageDataLoaderTests: XCTestCase {
         
         wait(for: [exp], timeout: 1.0)
     }
-    
-    
 }
